Update Relationships and strategies between Exercise and ExerciseSet
ExerciseSet:
Many-to-One relationship between ExerciseSet and Exercise, where exercise_id is the foreign key column in the ExerciseSet table.
changes in ExerciseSet cascade to its associated Exercise.

Exercise:
One-to-Many relationship between Exercise and ExerciseSet,
The mappedBy attribute indicates that the relationship is bidirectional, and it points to the exercise field in the ExerciseSet entity.
CascadeType.ALL means that changes in Exercise will cascade to its associated ExerciseSet instances.